Add main menu and light bar
To fill the light bar, you need to call the function created in the hud blueprint with a value between 0 and 1.
The "How To Play" menu is not finished. We just need to create the text.
The light bar needs to be added in the level (just copy the level blueprint of train level to instanciate the menu)